Code reorg.
